test LRUK

+++ src/include/buffer/lru_k_replacer.h
@@ -76,7 +76,7 @@ class LRUKReplacer {
   std::unordered_map<frame_id_t, LRUKNode> node_store_;
   size_t current_timestamp_{0};
   size_t curr_size_{0};
-  size_t replacer_size_;
+  int32_t replacer_size_;
   size_t k_;
   std::mutex latch_;
 };
